web/requirements.txt
psycopg2
Pillow
git+git://github.com/omab/python-social-auth.git@781173719e5e669e1306b8686941964c3ed92a5a
django>=1.10
xmltodict
djangorestframework
geocoder
geopy
markdown
django-filter
drf-nested-routers
django-pg-fts
jsonfield
django-jsonfield
django-debug-toolbar
admin-extra-urls
eventlet
celery==3.1.25
django-countries
django-extra-fields
django-activity-stream
git+https://github.com/celery/django-celery.git
thegamesdb
python-memcached
rest-framework-cache
dictdiffer
